+++ packages/neuron-wallet/src/block-sync-renderer/index.ts
@@ -1,63 +1,72 @@
-import { BrowserWindow } from 'electron'
 import path from 'path'
-import { fork } from 'child_process'
-import { Network, EMPTY_GENESIS_HASH } from 'models/network'
-import DataUpdateSubject from 'models/subjects/data-update'
-import env from 'env'
-import AddressCreatedSubject from 'models/subjects/address-created-subject'
-import WalletDeletedSubject from 'models/subjects/wallet-deleted-subject'
-import SyncedBlockNumber from 'models/synced-block-number'
+import { fork, ChildProcess } from 'child_process'
+import SyncApiController from 'controllers/sync-api'
 import NetworksService from 'services/networks'
 import AddressService from 'services/addresses'
 import WalletService from 'services/wallets'
-import logger from 'utils/logger'
-import { LumosCellQuery, LumosCell } from './sync/indexer-connector'
-import IndexerFolderManager from './sync/indexer-folder-manager'
-import { spawn, terminate, subscribe as subscribeToWorkerProcess } from 'utils/worker'
-import SyncApiController from 'controllers/sync-api'
-import type { SyncTask } from './task'
-import TxDbChangedSubject from 'models/subjects/tx-db-changed-subject'
-import AddressDbChangedSubject from 'models/subjects/address-db-changed-subject'
-import { queue } from 'async'
 import IndexerService from 'services/indexer'
+import { Network, EMPTY_GENESIS_HASH } from 'models/network'
+import DataUpdateSubject from 'models/subjects/data-update'
+import AddressCreatedSubject from 'models/subjects/address-created-subject'
+import WalletDeletedSubject from 'models/subjects/wallet-deleted-subject'
+import TxDbChangedSubject from 'models/subjects/tx-db-changed-subject'
+import { LumosCellQuery, LumosCell } from './sync/indexer-connector'
+import { WorkerMessage, StartParams, QueryIndexerParams } from './task'
+import logger from 'utils/logger'
+import CommonUtils from 'utils/common'
+import env from 'env'
 
-let syncTask: SyncTask | null
 let network: Network | null
+let child: ChildProcess | null = null
+let requestId = 0
+let requests = new Map<number, Record<'resolve' | 'reject', Function>>()
 
-const resetSyncTaskQueue = queue(async ({startTask, clearIndexerFolder}) => {
-  await WalletService.getInstance().maintainAddressesIfNecessary()
-  await killBlockSyncTask()
-  if (startTask) {
-    await createBlockSyncTask(clearIndexerFolder)
-  }
-}, 1)
+export const killBlockSyncTask = async () => {
+  const _child = child
+  child = null
+  if (!_child) { return }
 
-resetSyncTaskQueue.error(err => {
-  logger.error(err)
-})
+  logger.info('Sync:\tdrain requests')
+  await Promise.all(
+    [...requests.values()].map(
+      ({ reject }) =>
+        typeof reject === 'function'
+          ? reject()
+          : logger.error(`Worker:\treject is not a function, get ${reject}`)
+    ))
+    .finally(() => requests = new Map())
 
-export const resetSyncTask = async (startTask: boolean = true, clearIndexerFolder: boolean = false) => {
-  if (resetSyncTaskQueue.length() === 0) {
-    resetSyncTaskQueue.push({startTask, clearIndexerFolder})
-    await resetSyncTaskQueue.drain()
-  }
+  await waitForChildClose(_child)
+  await IndexerService.getInstance().stop()
 }
 
-if (BrowserWindow) {
-  AddressCreatedSubject.getSubject().subscribe(() => resetSyncTask())
-  WalletDeletedSubject.getSubject().subscribe(() => resetSyncTask())
+const waitForChildClose = (c: ChildProcess) => new Promise((resolve, reject) => {
+  c.once('close', resolve)
+  const msg: Required<WorkerMessage> = {
+    type: 'call',
+    id: requestId++,
+    channel: 'unmount',
+    message: null
+  }
+  c.send(msg, err => {
+    if (err) { reject(err) }
+  })
+}).catch(() => 0)
+
+export const resetSyncTask = async (startTask = true) => {
+  await killBlockSyncTask()
+
+  if (startTask) {
+    await WalletService.getInstance().maintainAddressesIfNecessary()
+    await CommonUtils.sleep(3000)
+    await createBlockSyncTask()
+  }
 }
 
 export const switchToNetwork = async (newNetwork: Network, reconnected = false, shouldSync = true) => {
-  const previousNetwork = network
-  network = newNetwork
+  if (!reconnected && network?.id === newNetwork.id) { return }
 
-  if (previousNetwork && !reconnected) {
-    if (previousNetwork.id === newNetwork.id) {
-      // There's no actual change. No need to reconnect.
-      return
-    }
-  }
+  network = newNetwork
 
   if (reconnected) {
     logger.info('Network:\treconnected to:', network)
@@ -68,50 +77,67 @@ export const switchToNetwork = async (newNetwork: Network, reconnected = false, 
   await resetSyncTask(shouldSync)
 }
 
-export const createBlockSyncTask = async (clearIndexerFolder: boolean) => {
-  const mercury = IndexerService.getInstance()
-  if (clearIndexerFolder) {
-    await new SyncedBlockNumber().setNextBlock(BigInt(0))
-    IndexerFolderManager.resetIndexerData()
-    await mercury.clearData()
-  }
+export const queryIndexer = async (query: LumosCellQuery): Promise<LumosCell[]> => {
+  const _child = child
+  if (!_child) { return [] }
+  const msg: Required<WorkerMessage<QueryIndexerParams>> = { type: 'call', id: requestId++, channel: 'queryIndexer', message: query }
+  return registerRequest(_child, msg).catch(err => {
+    logger.error(`Sync:\tfailed to register query indexer task`, err)
+    return []
+  }) as Promise<LumosCell[]>
+}
 
-  await mercury.start()
+export const createBlockSyncTask = async () => {
+  await IndexerService.getInstance().start()
 
-  logger.info('Sync:\tstarting background process')
+  logger.info('Sync:\tstart')
 
   // prevents the sync task from being started repeatedly if fork does not finish executing.
-  const childProcess = fork(path.join(__dirname, 'task-wrapper.js'), [], {
-    env: {
-      fileBasePath: env.fileBasePath
-    },
+  child = fork(path.join(__dirname, 'task-wrapper.js'), [], {
+    env: { fileBasePath: env.fileBasePath },
     stdio: ['ipc', process.stdout, 'pipe']
   })
 
-  childProcess.stderr!.setEncoding('utf8').on('data', data => {
-    logger.error('Sync:ChildProcess:', data)
-  })
-  syncTask = await spawn<SyncTask>(childProcess)
+  child.on('message', ({ id, message, channel }: WorkerMessage) => {
+    if (id !== undefined) {
+      if (!requests.has(id)) { return }
+      const { resolve } = requests.get(id)!
+      requests.delete(id)
+      if (typeof resolve === 'function') {
+        try {
+          resolve(message)
+        } catch (err) {
+          logger.error(`Sync:\t${err}`)
+        }
+      } else {
+        logger.error(`Sync:\tresolve expected, got ${resolve}`)
+      }
 
-  subscribeToWorkerProcess(syncTask, msg => {
-    switch (msg.channel) {
-      case 'cache-tip-block-updated':
-        SyncApiController.emiter.emit('cache-tip-block-updated', msg.result)
-        break
-      case 'tx-db-changed':
-        TxDbChangedSubject.getSubject().next(msg.result)
-        break
-      case 'address-db-changed':
-        AddressDbChangedSubject.getSubject().next(msg.result)
-        break
-      case 'wallet-deleted':
-      case 'address-created':
-      case 'indexer-error':
-        resetSyncTask()
-        break
-      default:
-        break
+    } else {
+      switch (channel) {
+        case 'cache-tip-block-updated':
+          SyncApiController.emiter.emit('cache-tip-block-updated', message)
+          break
+        case 'tx-db-changed':
+          TxDbChangedSubject.getSubject().next(message)
+          break
+        // case 'address-db-changed':
+        //   AddressDbChangedSubject.getSubject().next(msg.result)
+        //   break
+        case 'wallet-deleted':
+        case 'address-created':
+        case 'indexer-error':
+          resetSyncTask()
+          break
+        default:
+          break
+      }
+
     }
+  })
+
+  child.stderr!.setEncoding('utf8').on('data', async data => {
+    logger.error('Sync:ChildProcess:', data)
   })
 
 
@@ -124,33 +150,28 @@ export const createBlockSyncTask = async (clearIndexerFolder: boolean) => {
     actionType: 'update',
   })
 
+  const _child = child
+
   if (network.genesisHash !== EMPTY_GENESIS_HASH) {
-    syncTask!.start(
-      network.remote,
-      network.genesisHash,
-      await AddressService.getAddressesByAllWallets(),
-      IndexerService.LISTEN_URI
-    )
+    const addressMetas = await AddressService.getAddressesByAllWallets()
+    const message: StartParams = { genesisHash: network.genesisHash, url: network.remote, addressMetas, indexerUrl: IndexerService.LISTEN_URI }
+    const msg: Required<WorkerMessage<StartParams>> = { type: 'call', channel: 'start', id: requestId++, message }
+    return registerRequest(_child, msg).catch(err => {
+      logger.error(`Sync:\ffailed to register sync task`, err)
+    })
   }
 }
 
-export const killBlockSyncTask = async () => {
-  if (syncTask) {
-    logger.info('Sync:\tkill background process')
-    await syncTask.unmount()
-    await terminate(syncTask)
-    const mercury = IndexerService.getInstance()
-    await mercury.stop()
-    syncTask = null
-  }
-}
+export const registerRequest = (c: ChildProcess, msg: Required<WorkerMessage>) =>
+  new Promise((resolve, reject) => {
+    requests.set(msg.id, { resolve, reject })
+    c.send(msg, err => {
+      if (err) {
+        logger.error(`Sync:\tfailed to send message to child process: ${msg}`)
+        reject(err)
+      }
+    })
+  })
 
-export const queryIndexer = async (query: LumosCellQuery): Promise<LumosCell[]> => {
-  try {
-    const results = await syncTask?.queryIndexer(query) as LumosCell[]
-    return results || []
-  } catch (error) {
-    logger.error(error)
-    return []
-  }
-}
+AddressCreatedSubject.getSubject().subscribe(() => resetSyncTask())
+WalletDeletedSubject.getSubject().subscribe(() => resetSyncTask())