+++ packages/neuron-wallet/src/listeners/main/index.ts
@@ -1,12 +1,24 @@
-import { register as registerTxDbChanged } from './tx-db-changed'
-import { register as registerAddressDbChanged } from './address-db-changed'
-import ProcessUtils from 'utils/process'
 
-// register in app.ready
+import { debounceTime } from "rxjs/operators"
+import TxDbChangedSubject from "models/subjects/tx-db-changed-subject"
+import AddressDbChangedSubject from 'models/subjects/address-db-changed-subject'
+import DataUpdateSubject from "models/subjects/data-update"
+
+/**
+ * Update addresses and transactions actively
+ */
 export const register = () => {
-  if (ProcessUtils.isRenderer()) {
+  if (process.type === 'renderer') {
     throw new Error('Only call listeners/main in main process!')
   }
-  registerTxDbChanged()
-  registerAddressDbChanged()
+
+  TxDbChangedSubject.getSubject().pipe(debounceTime(500))
+    .subscribe(() => {
+      DataUpdateSubject.next({ dataType: 'transaction', actionType: 'update' })
+    })
+
+  AddressDbChangedSubject.getSubject().pipe(debounceTime(200))
+    .subscribe(() => {
+      DataUpdateSubject.next({ dataType: 'address', actionType: 'update' })
+    })
 }