+++ packages/neuron-wallet/src/services/indexer.ts
@@ -4,10 +4,14 @@ import fs from 'fs'
 import net from 'net'
 import { ChildProcess, spawn } from 'child_process'
 import process from 'process'
+import { dialog } from 'electron'
 import logger from 'utils/logger'
 import { Network } from 'models/network'
-import { deleteFolderRecursive } from 'block-sync-renderer/sync/indexer-folder-manager'
+import SyncedBlockNumber from 'models/synced-block-number'
 import NetworksService from './networks'
+import CommonUtils from 'utils/common'
+import { resetSyncTask } from 'block-sync-renderer'
+import { clean as cleanChain } from 'database/chain'
 
 const platform = (): string => {
   switch (process.platform) {
@@ -27,13 +31,13 @@ const { app } = env
 export default class IndexerService {
   private indexer: ChildProcess | null = null
   private static instance: IndexerService
-  private static indexerDataFolder = './ckb_indexer_data'
+  private static indexerDataFolder = 'ckb_indexer_data'
   public static PORT = '8118'
   public static get LISTEN_URI() {
     return `http://localhost:${IndexerService.PORT}`
   }
 
-  public static getInstance = () => {
+  static getInstance = () => {
     if (!IndexerService.instance) {
       IndexerService.instance = new IndexerService()
     }
@@ -42,70 +46,37 @@ export default class IndexerService {
 
   static getPath = (): string => {
     return app.isPackaged ?
-      path.join(path.dirname(app.getAppPath()), '..', './bin') :
+      path.join(path.dirname(app.getAppPath()), '..', 'bin') :
       path.join(__dirname, '../../bin',)
   }
 
   static getBinary = (): string => {
     const binary = app.isPackaged ?
-      path.resolve(IndexerService.getPath(), './ckb-indexer') :
-      path.resolve(IndexerService.getPath(), `./${platform()}`, './ckb-indexer')
+      path.resolve(IndexerService.getPath(), 'ckb-indexer') :
+      path.resolve(IndexerService.getPath(), `${platform()}`, 'ckb-indexer')
     return platform() === 'win' ? binary + '.exe' : binary
   }
 
-  getDataPath = (network: Network): string => {
-    return app.isPackaged ?
-      path.resolve(app.getPath('userData'), IndexerService.indexerDataFolder, './data', `./${network.genesisHash}`) :
-      path.resolve(app.getPath('userData'), './dev', IndexerService.indexerDataFolder, './data', `./${network.genesisHash}`)
-  }
-
-  async stop() {
-    return new Promise<void>(_resolve => {
-      const resolve = () => {
-        this.indexer?.removeAllListeners()
-        _resolve()
-      }
-      if (this.indexer) {
-        logger.info('Indexer:\tkilling')
-        this.indexer.once('close', () => resolve())
-        this.indexer.kill()
-        this.indexer = null
-      } else {
-        resolve()
-      }
-    })
-  }
-
-  start = async () => {
-    const network = NetworksService.getInstance().getCurrent()
-    await this.stop()
-    const dataPath = this.getDataPath(network)
-    IndexerService.createFolder(dataPath)
+  static ensurePortUsable = async () => {
+    const port = Number(IndexerService.PORT)
+    const isPortReachable = await IndexerService.isPortReachable(port)
+    if (!isPortReachable) {
+      return
+    }
+    IndexerService.PORT = (port + 1).toString()
     await IndexerService.ensurePortUsable()
-
-    this.indexer = spawn(IndexerService.getBinary(), ['-c', network.remote, '-s', dataPath, '-l', `127.0.0.1:${IndexerService.PORT}`])
-    logger.info(`Indexer:\tstart: PORT: ${IndexerService.PORT}...`)
-    this.indexer.stderr && this.indexer.stderr.on('data', data => {
-      logger.error('Indexer:\trun fail:', data.toString())
-      this.indexer = null
-    })
-
-    this.indexer.on('error', error => {
-      logger.error('Indexer:\trun fail:', error)
-      this.indexer = null
-    })
-
-    this.indexer.on('close', async () => {
-      logger.info('Indexer:\tprocess closed')
-      this.indexer = null
-    })
   }
 
-  clearData = async () => {
-    const network = NetworksService.getInstance().getCurrent()
-    await this.stop()
-    const dataPath = this.getDataPath(network)
-    deleteFolderRecursive(dataPath)
+  static clearCache = async (clearIndexerFolder = false) => {
+    await resetSyncTask(false)
+    await cleanChain()
+
+    if (clearIndexerFolder) {
+      IndexerService.getInstance().clearData()
+      await new SyncedBlockNumber().setNextBlock(BigInt(0))
+    }
+
+    await resetSyncTask(true)
   }
 
   static createFolder(dir: string) {
@@ -143,13 +114,77 @@ export default class IndexerService {
     }
   }
 
-  static ensurePortUsable = async () => {
-    const port = Number(IndexerService.PORT)
-    const isPortReachable = await IndexerService.isPortReachable(port)
-    if (!isPortReachable) {
+  stop = async () => {
+    const _indexer = this.indexer
+
+    if (!_indexer) { return }
+    logger.debug(`Indexer:\tstopping`)
+    // _indexer.removeAllListeners()
+    await new Promise((resolve, reject) => {
+      _indexer.once('close', code => {
+        return code ? reject(code) : resolve(code)
+      })
+      _indexer.kill()
+    }).catch(() => null)
+  }
+
+  start = async () => {
+    if (this.indexer) {
       return
     }
-    IndexerService.PORT = (port + 1).toString()
+    const network = NetworksService.getInstance().getCurrent()
+    const dataPath = this.#getDataPath(network)
+    IndexerService.createFolder(dataPath)
     await IndexerService.ensurePortUsable()
+
+    try {
+      const bin = IndexerService.getBinary()
+      const params = ['-c', network.remote, '-s', dataPath, '-l', `127.0.0.1:${IndexerService.PORT}`]
+      const indexer = spawn(bin, params)
+      this.indexer = indexer
+      logger.info(`Indexer:\tstart: PORT: ${IndexerService.PORT}...`)
+
+      indexer.stderr && indexer.stderr.on('data', async data => {
+        logger.debug(`Indexer stderr:\t`, data.toString())
+      })
+
+      indexer.on('error', (error: any) => {
+        logger.error('Indexer error:\t:', error.toString())
+        CommonUtils.sleep(3000)
+        this.start()
+      })
+
+      indexer.on('exit', () => {
+        logger.debug(`Indexer:\tstopped`)
+        this.indexer = null
+      })
+
+
+    } catch (err) {
+      logger.error(err)
+      dialog.showErrorBox("CKB Indexer", err.message)
+    }
+
+
+    // REFACTOR: use message to return the promise
+    while (!(await IndexerService.isPortReachable(+IndexerService.PORT))) {
+      await CommonUtils.sleep(1000)
+    }
+  }
+
+  clearData = () => {
+    const network = NetworksService.getInstance().getCurrent()
+    const dataPath = this.#getDataPath(network)
+    logger.debug(`Removing data ${dataPath}`)
+    fs.rmSync(dataPath, { recursive: true, force: true })
+
+    // remove legacy data
+    fs.rmSync(path.resolve(env.fileBasePath, 'indexer_data', network.genesisHash), { recursive: true, force: true })
+  }
+
+
+
+  #getDataPath = (network: Network): string => {
+    return path.resolve(env.fileBasePath, IndexerService.indexerDataFolder, 'data', `${network.genesisHash}`)
   }
 }