+++ packages/neuron-wallet/src/services/live-cell-service.ts
@@ -14,14 +14,14 @@ export default class LiveCellService {
     return LiveCellService.instance
   }
 
-  constructor() {}
+  constructor() { }
 
   private async getLiveCellsByScript(lock: Script | null, type: Script | null, data: string | null): Promise<LumosCell[]> {
     if (!lock && !type) {
       throw new Error('at least one parameter is required')
     }
 
-    const query: LumosCellQuery = {lock, type, data}
+    const query: LumosCellQuery = { lock, type, data }
     const liveCells: LumosCell[] = await queryIndexer(query)
     return liveCells
   }